% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_time_varying_functions.R
\name{add_list_periods}
\alias{add_list_periods}
\title{Helper function add lists of periods into single containing list which for
each age group has values for all periods}
\usage{
add_list_periods(l_l_period_def = NULL, l_period_def)
}
\arguments{
\item{l_l_period_def}{Existing list of lists.}

\item{l_period_add}{List of periods to add to the containing list, should
be defined using add_period calls.}
}
\value{
List (\code{l_l_period_def}) appended with \code{l_period_def} as a new entry.
}
\description{
\code{add_list_periods} adds lists of time periods as new entries in list of
age group lists.
}
