% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_time_varying_functions.R
\name{add_period}
\alias{add_period}
\title{Helper function add periods together into a list containing all periods}
\usage{
add_period(l_period_def = NULL, l_period_add)
}
\arguments{
\item{l_period_def}{Existing list of period definitions.}

\item{l_period_add}{Period to add to the full list, should be defined
using \code{make_period}.}
}
\value{
List (\code{l_period_def}) appended with \code{l_period_add} as a new entry.
}
\description{
\code{add_period} adds time periods as new entries in list of time periods
}
