% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_contact_matrix.R
\name{aggregate_pop_ages}
\alias{aggregate_pop_ages}
\title{Aggregate population ages into age groups of interest}
\usage{
aggregate_pop_ages(
  match_country = "",
  match_state = "",
  match_county = "",
  ages_to_cut = v_init_age_grps
)
}
\arguments{
\item{match_country}{Character that specifies which country population age
structure to return.}

\item{match_state}{Character that specifies which state population age
structure to return.}

\item{match_county}{Character that specifies which county population age
structure to return.}

\item{ages_to_cut}{Vector that specifies the age bins to aggregate and return.}
}
\description{
\code{aggregate_pop_ages} returns a data.frame with all of the
processed information on population age structure.
}
