% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_model_output_calc_plot_functions.R
\name{calc_basic_reproduction_number_kr}
\alias{calc_basic_reproduction_number_kr}
\title{Calculate basic reproduction number according to Keeling & Rohani}
\usage{
calc_basic_reproduction_number_kr(l_out_cosmo, v_time = 2:10)
}
\arguments{
\item{l_out_cosmo}{List with output from SC-COSMO and all parameters.}

\item{v_time}{Vector with times at which basic reproduction number is
desired to be calculated}
}
\value{
A scalar with the basic reproduction number.
}
\description{
\code{calc_basic_reproduction_number_kr} Calculates the basic reproduction
number according to Keeling & Rohani.
}
\references{
\enumerate{
\item
\item Emilia Vynnycky and Richard G White. An introduction to infectious
disease modelling. Oxford, New York: Oxford University Press, 2010.
\item Keeling, M. J., & Rohani, P. Modeling Infectious Diseases in
Humans and Animals. Princeton University Press, 2008.
}
}
