% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_model_output_calc_plot_functions.R
\name{calc_deathsdx_totals}
\alias{calc_deathsdx_totals}
\title{Total Covid-19 deaths from diagnosed infections}
\usage{
calc_deathsdx_totals(l_out_cosmo)
}
\arguments{
\item{l_out_cosmo}{List with output from SC-COSMO and all parameters.}
}
\value{
A data.frame with the total number of Covid-19 deaths from diagnosed
infections by age group and overall as columns over time.
}
\description{
\code{calc_deathsdx_totals} calculate total number of Covid-19 deaths
from diagnosed infections by age group and overall over time.
}
