% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_hospital_model_functions.R
\name{calc_dx_hosp}
\alias{calc_dx_hosp}
\title{Total Prevalent DX hospitalizations}
\usage{
calc_dx_hosp(l_hosp, l_out_cosmo)
}
\arguments{
\item{l_hosp}{List with output from the hospitalization simulation
(prep_hospitalizations)}

\item{l_out_cosmo}{List with output from SC-COSMO and all parameters}
}
\value{
A data.frame with the total number of total number of prevalent diagnosed
hospitalizations by age group and overall as columns over time.
}
\description{
Calculate total number of prevalent diagnosed hospitalizations.
}
