% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_hospital_model_functions.R
\name{calc_dx_hosp_nonicu}
\alias{calc_dx_hosp_nonicu}
\title{Total Prevalent DX Non-ICU hospitalizations}
\usage{
calc_dx_hosp_nonicu(l_hosp, l_out_cosmo)
}
\arguments{
\item{l_hosp}{List with output from the hospitalization simulation
(\code{prep_dx_hospitalizations}).}

\item{l_out_cosmo}{List with output from SC-COSMO and all parameters.}
}
\value{
A data.frame with the total number of total number of prevalent diagnosed
Non-ICU hospitalizations by age group and overall as columns over time.
}
\description{
\code{calc_dx_hosp_nonicu} calculate total number of prevalent diagnosed Non-ICU hospitalizations.
}
