% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_hospital_model_functions.R
\name{calc_dx_inchosp}
\alias{calc_dx_inchosp}
\title{Total Incident DX hospitalizations}
\usage{
calc_dx_inchosp(l_hosp, l_out_cosmo)
}
\arguments{
\item{l_hosp}{List with output from the hospitalization simulation
(\code{prep_dx_hospitalizations}).}

\item{l_out_cosmo}{List with output from SC-COSMO and all parameters.}
}
\value{
A data.frame with the total number of total number of incident diagnosed
hospitalizations by age group and overall as columns over time.
}
\description{
\code{calc_dx_inchosp} calculate total number of incident diagnosed hospitalizations.
}
