% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_model_output_calc_plot_functions.R
\name{calc_dxinc_totals}
\alias{calc_dxinc_totals}
\title{Total incident diagnosed infections (E and I)}
\usage{
calc_dxinc_totals(l_out_cosmo)
}
\arguments{
\item{l_out_cosmo}{List with output from SC-COSMO and all parameters}
}
\value{
A data.frame with the total number of incident diagnosed infections (E and I) by age group and overall
as columns over time
}
\description{
\code{calc_dxinc_totals} Calculate total number of incident diagnosed
infections (E and I) by age group and overall over time.
}
