% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_model_output_calc_plot_functions.R
\name{calc_idxcum_totals}
\alias{calc_idxcum_totals}
\title{Total cumulative diagnosed infectious infections}
\usage{
calc_idxcum_totals(l_out_cosmo)
}
\arguments{
\item{l_out_cosmo}{List with output from SC-COSMO and all parameters.}
}
\value{
A data.frame with the total number of cumulative diagnosed infectious infections by age group and overall
as columns over time.
}
\description{
\code{calc_idxcum_totals} calculate total number of cumulative diagnosed
infectious infections by age group and overall over time.
}
