% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_model_output_calc_plot_functions.R
\name{calc_incdeathsdx_totals}
\alias{calc_incdeathsdx_totals}
\title{Total Incident Covid-19 deaths from diagnosed infections}
\usage{
calc_incdeathsdx_totals(l_out_cosmo)
}
\arguments{
\item{l_out_cosmo}{List with output from SC-COSMO and all parameters.}
}
\value{
A data.frame with the total incident number of Covid-19 deaths from diagnosed
infections by age group and overall as columns over time.
}
\description{
\code{calc_incdeathsdx_totals} calculate total incident number of
Covid-19 deaths from diagnosed infections by age group and overall over time.
}
