% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_model_output_calc_plot_functions.R
\name{calc_popsize_totals}
\alias{calc_popsize_totals}
\title{Total population}
\usage{
calc_popsize_totals(l_out_cosmo)
}
\arguments{
\item{l_out_cosmo}{List with output from SC-COSMO and all parameters}
}
\value{
A data.frame with the total population by age group and overall
as columns over time
}
\description{
\code{calc_popsize_totals} Calculate total population by age group and
overall over time.
}
