% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_model_output_calc_plot_functions.R
\name{calc_reproduction_number_wt}
\alias{calc_reproduction_number_wt}
\title{Estimate the time dependent reproduction number according to Wallinga &
Teunis}
\usage{
calc_reproduction_number_wt(l_out_cosmo, v_time = 2:20, nsim_chosen = 200)
}
\arguments{
\item{l_out_cosmo}{List with output from SC-COSMO and all parameters.}

\item{v_time}{Vector with times at which the time-dependent reproduction
number is desired to be calculated}

\item{nsim_chosen}{number of simulations used by the WT algorithm to
compute confidence intervals which are often not used. Hence the number
is small. The default for WT in the package is 10,000 which uses a very
large amount of memory and is not appropriate for parallelization because
we will run out of RAM. The default is set to a much smaller number}
}
\value{
A data.frame with with the time-dependent reproduction number.
}
\description{
\code{calc_reproduction_number_wt} Calculates the time-dependent reproduction
number according to Wallinga & Teunis.
}
\references{
\enumerate{
\item
\item Wallinga, J., & Teunis, P. (2004). Different epidemic curves for severe
acute respiratory syndrome reveal similar impacts of control measures.
American Journal of Epidemiology, 160(6), 509–516.
https://doi.org/10.1093/aje/kwh255
}
}
