% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_model_output_calc_plot_functions.R
\name{calc_sx_totals}
\alias{calc_sx_totals}
\title{Total symptomatic infections (E and I)}
\usage{
calc_sx_totals(l_out_cosmo)
}
\arguments{
\item{l_out_cosmo}{List with output from SC-COSMO and all parameters.}
}
\value{
A data.frame with the total number of symptomatic infections (E and I)
by age group and overall as columns over time.
}
\description{
\code{calc_sx_totals} calculate total number of symptomatic
infections (E and I) by age group and overall over time.
}
