% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_decision_model_functions.R
\name{cosmo}
\alias{cosmo}
\title{The Stanford-CIDE COronavirus Simulation MOdel (SC-COSMO)}
\usage{
cosmo(l_params_all)
}
\arguments{
\item{l_params_all}{List with all parameters of decision model.}
}
\value{
A data.frame with population for each class by age group over time and a
list with all pareameters of decision model.
}
\description{
\code{cosmo} implements the Stanford-CIDE COronavirus Simulation MOdel
(SC-COSMO), which is an age-structure Susceptible-Exposed-Infectious-Recovered
(SEIR) model of the Coronavirus disease (Covid-19).
}
