% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_decision_model_functions.R
\name{cosmo_dXdt}
\alias{cosmo_dXdt}
\title{Derivatives of the Stanford-CIDE COronavirus Simulation MOdel (SC-COSMO)}
\usage{
cosmo_dXdt(time, v_pop, l_params_all)
}
\arguments{
\item{time}{Time at which the derivatives will be computed.}

\item{v_pop}{Vector with the population by class.}

\item{l_params_all}{List with all parameters of SC-COSMO model.}
}
\value{
A list with a vector of derivatives for each class by age group.
}
\description{
\code{cosmo_dXdt} computes the derivatives of the Stanford-CIDE COronavirus
Simulation MOdel (SC-COSMO).
}
