% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_decision_model_functions.R
\name{cosmo_dXdt_comp}
\alias{cosmo_dXdt_comp}
\title{Compiled Stanford-CIDE COronavirus Simulation MOdel (SC-COSMO) with detected
infectious states}
\usage{
cosmo_dXdt_comp(time, v_pop, l_params_all)
}
\description{
\code{cosmo_dXdt_comp} Compiled version of the Stanford-CIDE COronavirus
Simulation MOdel (SC-COSMO).
}
