% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_beta_mixing_interventions_functions.R
\name{gen_all_waifws}
\alias{gen_all_waifws}
\title{Efficiency by precomputing WAIFW matrices at each time step for
the Stanford-CIDE COronavirus Simulation MOdel (SC-COSMO)}
\usage{
gen_all_waifws(l_interventions, l_contact_info, v_names_ages, max_time)
}
\arguments{
\item{l_interventions}{List of interventions and timing.}

\item{l_contact_info}{List of WAIFW matrix inputs.}

\item{v_names_ages}{Vector of names for rows and columns of WAIFW matrices.}

\item{max_time}{Number of time steps to precompute WAIFW matrices for.}
}
\value{
A 3D matrix of appropriate WAIFW matrices for each time step in the model.
}
\description{
\code{gen_all_waifws} provides a way to precompute all WAIFW matrices
given the pattern of interventions at each time step for the
SC-COSMO model.
}
