% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_hosp_target_functions.R
\name{gen_hosp_targets}
\alias{gen_hosp_targets}
\title{Generate hospitalization targets.}
\usage{
gen_hosp_targets(
  n_time_stamp = "2020-12-21",
  n_date_ini = NULL,
  n_date_last = "2020-12-07"
)
}
\arguments{
\item{n_time_stamp}{Date at which hospitalization series were released.}

\item{n_date_ini}{Initial date to generate targets.}

\item{n_date_last}{Last date to generate targets.}
}
\value{
A data.frame with state-specific calibration targets.
}
\description{
\code{gen_hosp_targets} generates targets for hospitalization parameters.
}
