% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_decision_household_model_functions.R
\name{gen_household_transmission_mc_seir}
\alias{gen_household_transmission_mc_seir}
\title{Generate household transmission}
\usage{
gen_household_transmission_mc_seir(
  r_tau,
  n_hhsize,
  n_contacts_hh,
  v_HH,
  v_index_hh_sus,
  v_index_hh_exp,
  v_index_hh_inf,
  v_index_hh_rec,
  m_possibilities
)
}
\arguments{
\item{r_tau}{Transmission rate in the household model}

\item{n_hhsize}{Household size (integer)}

\item{n_contacts_hh}{Average number of within household contacts}

\item{v_HH}{State vector of within household epidemics}

\item{v_index_hh_sus}{Indexing column vector of susceptibles}

\item{v_index_hh_inf}{Indexing column vector of infectious}

\item{v_index_hh_rec}{Indexing column vector of recovered}

\item{m_possibilities}{Matrix with possible combinations of household
members within each epidemic compartment}
}
\value{
A scalar with the force of infection
}
\description{
\code{gen_household_transmission} generates household transmission (i.e.,
force of infection).
}
