% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_target_functions.R
\name{gen_targets}
\alias{gen_targets}
\title{Generate calibration targets}
\usage{
gen_targets(
  v_states_calib = "MCMA",
  n_time_stamp = "2020-09-13",
  n_date_ini = NULL,
  n_date_last = "2020-08-31"
)
}
\arguments{
\item{v_states_calib}{Vector of states used for calibration.}

\item{n_time_stamp}{Date at which Covid-19 series were released by the
Epidemiology Department of the Ministry of Health.}

\item{n_date_ini}{Initial date to generate targets.}

\item{n_date_last}{Last date to generate targets.}
}
\value{
A data.frame with state-specific calibration targets.
}
\description{
\code{gen_targets} generates state-specific calibration targets for
Mexico City Metropolitan Area based on symptomatic cases.
}
