% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_time_varying_functions.R
\name{gen_time_varying}
\alias{gen_time_varying}
\title{Helper function for creating time-varying vector of values}
\usage{
gen_time_varying(l_period_def, max_time)
}
\arguments{
\item{l_period_def}{Existing list of period definitions, created using
\code{add_period}.}

\item{max_time}{Number of days that should be estimated, will fill if not
fully covered by final stop in \code{l_period_def}.}
}
\value{
Vector of time-varying parameter values at minimum of length max_time
(could be longer).
}
\description{
\code{gen_time_varying} creates a time-varying vector of values based on
entries in \code{l_period_def}.
}
