% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_time_varying_functions.R
\name{general_logit}
\alias{general_logit}
\title{Generalized logistic function}
\usage{
general_logit(logit_lb, logit_ub, logit_change_rate, logit_change_mid, logit_t)
}
\arguments{
\item{logit_lb}{lower asymptote}

\item{logit_ub}{upper asymptote}

\item{logit_change_rate}{rate of change from lower to upper asymptote}

\item{logit_change_mid}{point in time (t) where function has gone halfway from lb to ub}

\item{logit_t}{time point(s) at which function should be evaluated}
}
\value{
value(s) of generalized logit function | inputs
}
\description{
\code{general_logit} Computes a generalized logistic function whose starting (lb)
and ending values (ub) are determined by parameters. Rate of change from lb to ub
and timing of change are also controled by parameters.
}
