% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_calibration_functions.R
\name{get_bounds}
\alias{get_bounds}
\title{Get bounds for parameters to calibrate}
\usage{
get_bounds()
}
\value{
A list with lower and upper bounds and their standard errors
}
\description{
\code{get_bounds} defines the bounds for each of the calibrated parameters.
}
