% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_case_detection_functions.R
\name{get_case_detection}
\alias{get_case_detection}
\title{Case detection parameters}
\usage{
get_case_detection(time, l_params_all, severity)
}
\arguments{
\item{time}{Time (numeric, in days) at which case detection rate is evaluated.}

\item{l_params_all}{List with all parameters of decision model.
Case detection rates should be stored as a list within
\code{l_params_all$l_case_detection}.}

\item{severity}{Severity level and state (exp1, exp2, inf1, inf2).}
}
\value{
Case detection rate for a given severity (1, 2) or state (exp, inf) at a
given point in time.
}
\description{
\code{get_case_detection} pulls case detection parameters for a given time.
}
