% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_mortality_functions.R
\name{get_cfr}
\alias{get_cfr}
\title{Proportion (age-specific) of detected cases that die}
\usage{
get_cfr(time, l_params_all)
}
\arguments{
\item{time}{Time (numeric, in days) at which \code{v_cfr} is evaluated.}

\item{l_params_all}{List with all parameters of decision model.}
}
\value{
Proportion (age-specific) of detected cases that die at a given point in time.
}
\description{
\code{get_cfr} pulls age-specific fraction of detected cases that die at a
given time.
}
