% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_contact_matrix.R
\name{get_contact_matrices_world}
\alias{get_contact_matrices_world}
\title{Get contact matrices}
\usage{
get_contact_matrices_world(reload = FALSE)
}
\arguments{
\item{reload}{Flag (default is FALSE) of whether to
redownload and process the data file.}
}
\description{
\code{get_contact_matrices_world} returns a data.frame with all of the
processed contact matrices for all countries.
}
