% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_contact_matrix.R
\name{get_contact_matrix}
\alias{get_contact_matrix}
\title{Get contact matrix main function}
\usage{
get_contact_matrix(
  v_init_age_grps = c(0, 5, 15, 25, 45, 55, 65, 70),
  country = "",
  state = "",
  county = "",
  src_country = "",
  density = "",
  v_custom_prop_pop = "",
  recache = FALSE
)
}
\arguments{
\item{v_init_age_grps}{Vector that specifies the age bins to aggregate and
return.}

\item{country}{Country of desired data.}

\item{state}{State of desired data.}

\item{county}{County of desired data.}

\item{src_country}{Country to use for contact matrix structure if country
is not in data.frame.}

\item{density}{Population density of country/state/county.}

\item{v_custom_prop_pop}{Vector that specifies the proportion of total
population for each age group.}

\item{recache}{Flag (default is FALSE) of whether to run function and replace
existing cache. If FALSE, will check for cache and use if cached matrix exists.}
}
\value{
A list with overall and setting specific contact matrices.
}
\description{
\code{get_contact_matrix} wrapper function that allows user to
subset data from various sources.
If only county is specified (other parameters = ""), it provides national
level data.
If county is also specified (and the dataset contains counties), then only
those counties within the state(s) are returned.
}
