% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_densities.R
\name{get_countrycode}
\alias{get_countrycode}
\title{Get a three letter code for a country's name}
\usage{
get_countrycode(country)
}
\description{
\code{get_countrycode} returns a string with a 3 letter code
for a country's name to help with matching between datasets.
}
