% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_cumulative_mortality_risk.R
\name{get_cum_mort_risk}
\alias{get_cum_mort_risk}
\title{Generate cumulative mortality risk}
\usage{
get_cum_mort_risk(save_data = FALSE)
}
\arguments{
\item{save_data}{Flag (default is FALSE) of whether to save data as a csv file.}
}
\value{
A data.frame with cumulative mortality risk expected by 30 and 60 days.
}
\description{
\code{get_cum_mort_risk} generates a 30- and 60-day cumulative mortality risk
of Covid-19 for MCMA.
}
