% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_densities.R
\name{get_densities}
\alias{get_densities}
\title{Get population density}
\usage{
get_densities(country = "", state = "", county = "")
}
\arguments{
\item{country}{Country of desired data.}

\item{state}{State of desired data.}

\item{county}{County of desired data.}

\item{reload}{Flag (default is FALSE) of whether to redownload and process
the data file.}
}
\value{
A data.frame with population density of desired country/state/county.
}
\description{
\code{get_densities} wrapper function that allows user to subset data from
various sources.
If only county is specified (other parameters = ""), it provides
national level data.
Currently for US, if state(s) is specified, then it returns
lowest geographical level (state or county) that it has
within those specified state(s).
If county is also specified (and the dataset contains counties), then only
those counties within the state(s) are returned.
If county is also specified (and the dataset contains counties), then only
those counties within the state(s) are returned.
}
