% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_densities.R
\name{get_densities_mx}
\alias{get_densities_mx}
\title{Get population density for Mexico subnational regions}
\usage{
get_densities_mx(reload = FALSE)
}
\arguments{
\item{reload}{Flag (default is FALSE) of whether to
redownload and process the data file.}
}
\description{
\code{get_densities_mx} returns a data.frame with all of the
processed information on weighted population density.
}
