% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_densities.R
\name{get_fipscodes_mx}
\alias{get_fipscodes_mx}
\title{Get FIPS codes for Mexican states}
\usage{
get_fipscodes_mx(state)
}
\arguments{
\item{state}{State of desired data.}
}
\description{
\code{get_fipscodes_mx} returns a string with the FIPS code
for Mexican states.
}
