% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_hosp_calibration_functions.R
\name{get_hosp_bounds}
\alias{get_hosp_bounds}
\title{Get bounds for hospitalization parameters to calibrate}
\usage{
get_hosp_bounds()
}
\value{
A list with lower and upper bounds and their standard errors.
}
\description{
\code{get_hosp_bounds} defines the bounds for each of the calibrated
hospitalization parameters.
}
