% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_cases_functions.R
\name{get_idx_scale_factor}
\alias{get_idx_scale_factor}
\title{Get reduction factor on transmission from detected infectious (IDX)}
\usage{
get_idx_scale_factor(time, l_params_all)
}
\arguments{
\item{time}{the time (numeric, in days) at which v_idx_scale_factor is
evaluated}

\item{l_params_all}{List with all parameters of decision model.}
}
\value{
Reduction (age-specific) on infectiousness of detected cases at a given point
in time.
}
\description{
\code{get_idx_scale_factor} pulls age-specific reduction factor on
on transmission from detected infectious
}
