% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_lifetables.R
\name{get_lifetables}
\alias{get_lifetables}
\title{Get lifetables}
\usage{
get_lifetables(country = "", state = "", county = "")
}
\arguments{
\item{country}{Country of desired data.}

\item{state}{State of desired data.}

\item{county}{County of desired data.}
}
\value{
A data.frame with all of the processed information on lifetables.
}
\description{
\code{get_lifetables} wrapper function that allows user to
subset data from various sources.
If only county is specified (other parameters = ""),
it provides national level data.
If county is also specified (and the dataset contains counties),
then only those counties within the state(s) are returned.
}
