% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_lifetables.R
\name{get_lifetables_mx}
\alias{get_lifetables_mx}
\title{Get lifetables for MX states}
\usage{
get_lifetables_mx(reload = FALSE)
}
\arguments{
\item{reload}{Flag (default is FALSE) of whether to
redownload and process the data file}
}
\description{
\code{get_lifetables_mx} returns a data.frame with all of the
processed information on lifetables.
}
