% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_hosp_prop_functions.R
\name{get_MCMA_prop_hosp}
\alias{get_MCMA_prop_hosp}
\title{Generate hospitalization proportions.}
\usage{
get_MCMA_prop_hosp(
  n_date_ini = "2020-02-24",
  n_date_last = "2020-12-07",
  save_data = FALSE
)
}
\arguments{
\item{n_date_ini}{Character specifying the initial date to compute proportions.}

\item{n_date_last}{Character specifying the last date to compute proportions.}

\item{save_data}{Flag (default is FALSE) of whether to save the data as a
csv file.}
}
\value{
A data.frame with proportion of incident detected cases hospitalized.
}
\description{
\code{get_MCMA_prop_hosp} wrangles individual Covid-19 data for MCMA and
calculates the proportion of incident detected cases hospitalized.
}
