% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_time_varying_functions.R
\name{get_non_const_multiage_list}
\alias{get_non_const_multiage_list}
\title{Helper convenience function that defines age-specific lists of periods for values that
change over the duration of the simulation.}
\usage{
get_non_const_multiage_list(v_time_stop, m_ageval)
}
\arguments{
\item{v_time_stop}{Vector with time period at which values change (in days).
Last entry should be \code{n_t}.}

\item{m_ageval}{Matrix of values for each age group (\code{n_ages}) at each
time period.}
}
\value{
List (\code{l_l_period}) list of lists of periods (1 per age group) with
non-constant values.
}
\description{
\code{get_non_const_multiage_list} adds lists of time periods as new entries
in list of age group lists.
}
