% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_population_ages.R
\name{get_population_ages}
\alias{get_population_ages}
\title{Get population and age structure}
\usage{
get_population_ages(country = "", state = "", county = "")
}
\arguments{
\item{country}{Country of desired data.}

\item{state}{State of desired data.}

\item{county}{County of desired data.}
}
\value{
A data.frame with all of the processed information on population age structure.
}
\description{
\code{get_population_ages} wrapper function that allows user to
subset data from various sources.
If only county is specified (other parameters = ""),
it provides national level data.
If county is also specified (and the dataset contains counties),
then only those counties within the state(s) are returned.
}
