% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_population_ages.R
\name{get_population_ages_mx}
\alias{get_population_ages_mx}
\title{Get population and age structure for MX subnational regions}
\usage{
get_population_ages_mx(reload = FALSE)
}
\arguments{
\item{reload}{Flag (default is FALSE) of whether to redownload and process
the data file}
}
\description{
\code{get_population_ages_mx} returns a data.frame with all of the
processed information on population age structure.
}
