% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_population_birth.R
\name{get_population_birth}
\alias{get_population_birth}
\title{Get crude birth rates}
\usage{
get_population_birth(country = "", state = "", county = "")
}
\arguments{
\item{country}{Country of desired data.}

\item{state}{State of desired data.}

\item{county}{County of desired data.}
}
\value{
A data.frame with all of the processed information on crude birth rate.
}
\description{
\code{get_population_birth} wrapper function that allows user to
subset data from various sources.
If only county is specified (other parameters = ""),
it provides national level data.
If county is also specified (and the dataset contains counties),
then only those counties within the state(s) are returned.
}
