% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_population_birth.R
\name{get_population_birth_mx}
\alias{get_population_birth_mx}
\title{Get crude birth rates for MX states}
\usage{
get_population_birth_mx(reload = FALSE)
}
\arguments{
\item{reload}{Flag (default is FALSE) of whether to redownload and process
the data file.}
}
\description{
\code{get_population_birth_mx} returns a data.frame with all of the
processed information on crude birth rate.
}
