% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_projections_functions.R
\name{get_projection_scenarios}
\alias{get_projection_scenarios}
\title{Generates interventions input list for performing projections
with the SC-COSMO model}
\usage{
get_projection_scenarios(
  n_t,
  v_soc_dist_factor,
  v_mean_soc_dist_factor,
  v_n_date0_NPI,
  date_proj0
)
}
\arguments{
\item{n_t}{Simulation total time.}

\item{v_soc_dist_factor}{Vector with social distancing multipliers for the
different mobility segments calibrated to.}

\item{v_mean_soc_dist_factor}{Vector with mean social distancing multipliers
for different mobility segments calibrated to.}

\item{v_n_date0_NPI}{Vector with the time steps (0 = \code{date_init}) at
which effect of NPI changed in the calibration period.}

\item{date_proj0}{the time step (0 = \code{date_init}) where projection starts
(calibration period is over).}
}
\value{
A list of named (scenarios) of intervention lists formatted for
input into the SC-COSMO model.
}
\description{
\code{get_projection_scenarios} generates interventions input
list for performing projections with the SC-COSMO model
for selected state.
}
