% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_hospital_model_functions.R
\name{get_prop_hosp}
\alias{get_prop_hosp}
\title{Get proportion of cases that go into hospitalization over time}
\usage{
get_prop_hosp(time, l_params_all)
}
\arguments{
\item{time}{the time (numeric, in days) at which p_hosp is evaluated}

\item{l_params_all}{List with all parameters of decision model.}
}
\value{
Proportion hospitalized (age-specific) at a given point in time.
}
\description{
\code{get_prop_hosp} pulls age-specific proportion of incident detected cases
hospitalized for a given time
}
