% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_beta_mixing_interventions_functions.R
\name{get_waifw}
\alias{get_waifw}
\title{Gets correct mixing matrices for
The Stanford-CIDE COronavirus Simulation MOdel (SC-COSMO)}
\usage{
get_waifw(time, l_params_all)
}
\arguments{
\item{time}{current time model is evaluating}

\item{l_params_all}{List with all parameters of decision model}
}
\value{
an overall WAIFW matrix appropriate for the model given the
current time and interventions being used
}
\description{
\code{get_waifw} provides a wrapper to determine which WAIFW matrices
are the correct ones given intervention timings.
}
