% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_hosp_target_functions.R
\name{hosp_acor2cov}
\alias{hosp_acor2cov}
\title{Covariance matrix from autocorrelation coefficients and standard
deviations (or standard errors)}
\usage{
hosp_acor2cov(v_acor, v_sd)
}
\arguments{
\item{v_acor}{Vector of autocorrelation coefficients indexed by time}

\item{v_sd}{Vector of standard deviations (or standard errors)}
}
\value{
A covariance matrix.
}
\description{
\code{hosp_acor2cov} computes covariance matrix from autocorrelation coefficients
and standard deviations (or standard errors) of means.
}
