% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_hosp_calibration_functions.R
\name{hosp_calibration_out}
\alias{hosp_calibration_out}
\title{Generate model outputs for calibration from a hospitalization parameter set}
\usage{
hosp_calibration_out(
  v_params_calib,
  l_params_all,
  n_lag_inf = NULL,
  n_lag_conf = 12,
  l_dates_hosp_targets
)
}
\arguments{
\item{v_params_calib}{Vector of hospitalization parameters that need to be
calibrated.}

\item{l_params_all}{List with all parameters of the decision model.}

\item{n_lag_inf}{Lag in time series of infectious individuals.}

\item{n_lag_conf}{Lag in time series to account for the difference from
first symptomatic.}

\item{l_dates_hosp_targets}{List of initial and final date of target series.}
}
\value{
A list with model-predicted outcomes for each target.
}
\description{
\code{hosp_calibration_out} computes model hospitalization outputs
to be used for calibration routines.
}
