% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_hosp_calibration_functions.R
\name{hosp_likelihood}
\alias{hosp_likelihood}
\title{Likelihood}
\usage{
hosp_likelihood(v_params)
}
\arguments{
\item{v_params}{Vector (or matrix) of hospitalization parameters.}
}
\value{
A scalar (or vector) with likelihood values.
}
\description{
\code{hosp_likelihood} computes a likelihood value for one (or multiple)
hospitalization parameter set(s).
}
