% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_hosp_calibration_functions.R
\name{hosp_log_lik_opt}
\alias{hosp_log_lik_opt}
\title{Log-likelihood function for a hospitalization parameter set}
\usage{
hosp_log_lik_opt(v_params, ...)
}
\arguments{
\item{v_params}{Vector (or matrix) of hospitalization parameters.}

\item{...}{Further arguments to be passed to.}
}
\value{
A scalar (or vector) with log-likelihood values.
}
\description{
\code{hosp_log_lik_opt} computes a log-likelihood value for one (or multiple)
hospitalization parameter set(s) to be used in optimization procedures.
}
