% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_hosp_calibration_functions.R
\name{hosp_log_lik_par}
\alias{hosp_log_lik_par}
\title{Parallel evaluation of log-likelihood function for a sets of hospitalization
parameters}
\usage{
hosp_log_lik_par(v_params, log_lik_offset, ...)
}
\arguments{
\item{v_params}{Vector (or matrix) of model parameters.}

\item{log_lik_offset}{Offset applied to log-likelihood values.}

\item{...}{Further arguments to be passed to.}
}
\value{
A scalar (or vector) with log-likelihood values.
}
\description{
\code{hosp_log_lik_par} computes a log-likelihood value for one (or multiple)
hospitalization parameter set(s) using parallel computation.
}
