% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_hosp_calibration_functions.R
\name{hosp_log_post}
\alias{hosp_log_post}
\title{Evaluate log-posterior of calibrated hospitalization parameters}
\usage{
hosp_log_post(v_params, ...)
}
\arguments{
\item{v_params}{Vector (or matrix) of model parameters.}

\item{...}{Further arguments to be passed to.}
}
\value{
A scalar (or vector) with log-posterior values.
}
\description{
\code{hosp_log_post} computes a log-posterior value for one (or multiple)
hospitalization parameter set(s) based on the simulation model,
likelihood functions and prior distributions.
}
