% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_hosp_calibration_functions.R
\name{hosp_log_post_opt}
\alias{hosp_log_post_opt}
\title{Evaluate log-posterior of calibrated hospitalization parameters for OPTIMIZATION purposes}
\usage{
hosp_log_post_opt(v_params, ...)
}
\arguments{
\item{v_params}{Vector (or matrix) of hospitalization parameters.}

\item{...}{Further arguments to be passed to.}
}
\value{
A scalar (or vector) with log-posterior values.
}
\description{
\code{hosp_log_post_opt} computes a log-posterior value for one (or multiple)
hospitalization parameter set(s) based on the simulation model, likelihood
functions and prior distributions. Used for OPTIMIZATION purposes,
NOT Bayesian estimation.
}
