% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_hosp_calibration_functions.R
\name{hosp_log_prior}
\alias{hosp_log_prior}
\title{Evaluate log-prior of calibrated hospitalization parameters}
\usage{
hosp_log_prior(
  v_params,
  v_param_names = names(get_hosp_bounds()$v_lb),
  v_lb = get_hosp_bounds()$v_lb,
  v_ub = get_hosp_bounds()$v_ub
)
}
\arguments{
\item{v_params}{Vector (or matrix) of model parameters.}

\item{v_param_names}{Vector with parameter names.}

\item{v_lb}{Vector with upper bounds for each parameter.}

\item{v_ub}{Vector with lower bounds for each parameter.}
}
\value{
A scalar (or vector) with log-prior values.
}
\description{
\code{hosp_log_prior} computes a log-prior value for one (or multiple)
hospitalization parameter set(s) based on their prior distributions.
}
