% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_hosp_calibration_functions.R
\name{hosp_posterior}
\alias{hosp_posterior}
\title{Evaluate posterior of calibrated hospitalization parameters}
\usage{
hosp_posterior(v_params, ...)
}
\arguments{
\item{v_params}{Vector (or matrix) of model parameters.}

\item{...}{Further arguments to be passed to.}
}
\value{
A scalar (or vector) with posterior values.
}
\description{
\code{hosp_posterior} computes a posterior value for one (or multiple) hospitalization
parameter set(s).
}
