% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_hosp_calibration_functions.R
\name{hosp_prior}
\alias{hosp_prior}
\title{Evaluate prior of calibrated hospitalization parameters}
\usage{
hosp_prior(v_params)
}
\arguments{
\item{v_params}{Vector (or matrix) of model parameters.}
}
\value{
A scalar (or vector) with prior values.
}
\description{
\code{hosp_prior} computes a prior value for one (or multiple) hospitalization
parameter set(s).
}
