% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_calibration_functions.R
\name{likelihood}
\alias{likelihood}
\title{Likelihood}
\usage{
likelihood(v_params)
}
\arguments{
\item{v_params}{Vector (or matrix) of model parameters.}
}
\value{
A scalar (or vector) with likelihood values.
}
\description{
\code{likelihood} computes a likelihood value for one (or multiple)
parameter set(s).
}
