% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_calibration_functions.R
\name{log_lik}
\alias{log_lik}
\title{Log-likelihood function for a parameter set}
\usage{
log_lik(v_params, ...)
}
\arguments{
\item{v_params}{Vector (or matrix) of model parameters.}

\item{...}{Further arguments to be passed to.}
}
\value{
A scalar (or vector) with log-likelihood values.
}
\description{
\code{log_lik} computes a log-likelihood value for one (or multiple)
parameter set(s).
}
