% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_calibration_functions.R
\name{log_lik_par}
\alias{log_lik_par}
\title{Parallel evaluation of log-likelihood function for a sets of parameters}
\usage{
log_lik_par(v_params, log_lik_offset, ...)
}
\arguments{
\item{v_params}{Vector (or matrix) of model parameters.}

\item{log_lik_offset}{Offset applied to log-likelihood values.}

\item{...}{Further arguments to be passed to.}
}
\value{
A scalar (or vector) with log-likelihood values.
}
\description{
\code{log_lik_par} computes a log-likelihood value for one (or multiple)
parameter set(s) using parallel computation.
}
