% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_calibration_functions.R
\name{log_prior_opt}
\alias{log_prior_opt}
\title{Evaluate log-prior of calibrated parameters for OPTIMIZATION purposes}
\usage{
log_prior_opt(
  v_params,
  v_param_names = names(get_bounds()$v_lb),
  v_lb = get_bounds()$v_lb,
  v_ub = get_bounds()$v_ub
)
}
\arguments{
\item{v_params}{Vector (or matrix) of model parameters.}

\item{v_param_names}{Vector with parameter names.}

\item{v_lb}{Vector with upper bounds for each parameter.}

\item{v_ub}{Vector with lower bounds for each parameter.}
}
\value{
A scalar (or vector) with log-prior values.
}
\description{
\code{log_prior} computes a log-prior value for one (or multiple) parameter
set(s) based on their prior distributions. Used for OPTIMIZATION
purposes, NOT Bayesian estimation.
}
