% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_time_varying_functions.R
\name{m_general_logit}
\alias{m_general_logit}
\title{Generalized logistic function (matrix version)}
\usage{
m_general_logit(
  v_logit_lb,
  v_logit_ub,
  v_logit_change_rate,
  v_logit_change_mid,
  logit_t
)
}
\arguments{
\item{v_logit_lb}{Age-specific lower asymptote}

\item{v_logit_ub}{Age-specific upper asymptote}

\item{v_logit_change_rate}{Age-specific rate of change from lower to upper
asymptote}

\item{v_logit_change_mid}{Age-specific point in time (t) where function has
gone halfway from lb to ub}

\item{logit_t}{time point(s) at which function should be evaluated}
}
\value{
Matrix with value(s) of generalized logit function | inputs
}
\description{
\code{m_general_logit} Computes a generalized logistic function whose
starting (lb) and ending values (ub) are determined by parameters.
Rate of change from lb to ub and timing of change are also controled by
parameters.
}
