% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_intervention_construction_functions.R
\name{make_intervention}
\alias{make_intervention}
\title{Constructing intervention descriptions for each intervention in the
the Stanford-CIDE COronavirus Simulation MOdel (SC-COSMO)}
\usage{
make_intervention(intervention_type, time_start, time_stop, ...)
}
\arguments{
\item{intervention_type}{One of the currently implemented intervention
types in the SC-COSMO model: 1) "StatusQuo"; 2) "SocialDistancing";
3) "SocialDistancingLinear".}

\item{time_start}{Days since model start when intervention begins.}

\item{time_stop}{Days since model start when intervention ends.}

\item{...}{Named values specific to a given intervention type.}
}
\value{
A list object containing appropriately named values for intervention
objects.
}
\description{
\code{make_intervention} provides a generic constructor to build
intervention "objects".
}
