% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_time_varying_functions.R
\name{make_period}
\alias{make_period}
\title{Helper function to define a time chunk for a time varying parameter}
\usage{
make_period(functional_form, time_start, time_stop, val_start, val_end, ...)
}
\arguments{
\item{functional_form}{character string of "linear", "constant", or "general_logit"}

\item{time_start}{numeric time period start (in days)}

\item{time_stop}{numeric time period end (in days)}

\item{val_start}{starting value for time varying parameter at time_start}

\item{val_end}{ending value for time varying parameter at time_end}
}
\value{
list with parameters initialized
}
\description{
\code{make_period} Defines a time chunk for a generic time varying parameter.
Includes some required arguments, but also allows additional arguments.
}
