% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_model_inputs_functions.R
\name{obtain_init_pop}
\alias{obtain_init_pop}
\title{Obtain initial population}
\usage{
obtain_init_pop(
  v_init_age_grps = c(0, 5, 15, 25, 45, 55, 65, 70),
  v_inf_init_ages,
  v_names_ages,
  n_ages,
  v_names_states,
  n_states,
  ctry = "Italy",
  ste = NULL,
  cty = NULL,
  year = 2020
)
}
\arguments{
\item{v_init_age_grps}{Vector with initial ages of age groups.}

\item{v_inf_init_ages}{Vector with number of individuals in first infectious
class for each age group.}

\item{v_names_ages}{Vector with age groups.}

\item{n_ages}{Number of age groups.}

\item{v_names_states}{Vector with names of compartment states.}

\item{n_states}{Number of compartment states.}

\item{ctry}{Country.}

\item{ste}{State.}

\item{cty}{County or municipality.}

\item{year}{Year.}
}
\value{
An array with the initial population at each class by age group.
}
\description{
\code{obtain_init_pop} generates the number of people at each compartment by
age group.
}
