% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_model_inputs_functions.R
\name{obtain_setting_data}
\alias{obtain_setting_data}
\title{Obtain initial population}
\usage{
obtain_setting_data(
  v_init_age_grps,
  ctry = "Mexico",
  ste = "Mexico City",
  cty = "Mexico City",
  year = 2020
)
}
\arguments{
\item{v_init_age_grps}{Vector with initial ages of age groups.}

\item{ctry}{Country.}

\item{ste}{State.}

\item{cty}{County or municipality.}

\item{year}{Year.}
}
\value{
An array with the initial population at each class by age group.
}
\description{
\code{obtain_setting_data} generates the number of people at each compartment by
age group.
}
