% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_model_output_calc_plot_functions.R
\name{plot_cosmo_data}
\alias{plot_cosmo_data}
\title{Plot COSMO data}
\usage{
plot_cosmo_data(
  df_plotting,
  proportion = FALSE,
  age_proportion = FALSE,
  only_all = TRUE,
  print_plot = TRUE,
  label_y
)
}
\arguments{
\item{df_plotting}{Data.frame of data to be plotted.}

\item{proportion}{Flag (default is FALSE) of whether to divide the outcome
of interest by total population.}

\item{age_proportion}{Flag (default is FALSE) of whether to divide
the outcome of interest by age-specific total population.}

\item{only_all}{Flag (default is TRUE) of whether only overall ("All") data
should be plotted.}

\item{print_plot}{Flag (default is TRUE) of whether the plot should be printed
in screen.}

\item{label_y}{y-axis label.}
}
\value{
A ggplot object.
}
\description{
\code{plot_cosmo_data} a generic function called by various wrapper
functions to plot a data frame returned by one of our \code{calc} methods by
age group and/or in total over time.
}
