% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_model_output_calc_plot_functions.R
\name{plot_dxinc_totals}
\alias{plot_dxinc_totals}
\title{Plot total incident diagnosed cases (E and I)}
\usage{
plot_dxinc_totals(
  l_out_cosmo,
  proportion = FALSE,
  age_proportion = FALSE,
  only_all = TRUE,
  print_plot = TRUE
)
}
\arguments{
\item{l_out_cosmo}{List with output from SC-COSMO and all parameters}

\item{proportion}{Logical. Divides incident diagnosed cases by total population}

\item{age_proportion}{Logical. Divides incident diagnosed cases by age-specific total population}

\item{only_all}{logical. Should only overall total cases be plotted}
}
\value{
A data.frame with the total number of incident diagnosed cases (E and I) by age group and in total
as columns over time
}
\description{
\code{plot_dxinc_totals} Plot total number of incident diagnosed
cases (E and I) by age group and/or in total over time.
}
