% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_hosp_prop_functions.R
\name{plot_hosp_prop}
\alias{plot_hosp_prop}
\title{Plot proportion of incident detected cases hospitalized}
\usage{
plot_hosp_prop(df_hosp_prop, save_plot = FALSE)
}
\arguments{
\item{df_hosp_prop}{Data.frame of data of observed and estimated proportion
of incident detected cases hospitalized to be plotted.}

\item{save_plot}{Flag (default is FALSE) of whether to save the plot.}
}
\value{
A ggplot object.
}
\description{
\code{plot_hosp_prop} plot observed and estimated proportion of incident
detected cases hospitalized.
}
