% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_hosp_target_functions.R
\name{plot_hosp_targets}
\alias{plot_hosp_targets}
\title{Plot hospitalization targets}
\usage{
plot_hosp_targets(l_hosp_targets, print_plot = TRUE, save_plot = TRUE)
}
\arguments{
\item{l_hosp_targets}{List of calibration targets.}

\item{print_plot}{Flag (default is FALSE) of whether to print plot.}

\item{save_plot}{Flag (default is FALSE) of whether to save plot.}
}
\value{
A ggplot object.
}
\description{
\code{plot_hosp_targets} plot calibration targets.
}
