% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_model_output_calc_plot_functions.R
\name{plot_infinc_totals}
\alias{plot_infinc_totals}
\title{Plot total incident infections}
\usage{
plot_infinc_totals(
  l_out_cosmo,
  proportion = FALSE,
  age_proportion = FALSE,
  only_all = TRUE,
  print_plot = TRUE
)
}
\arguments{
\item{l_out_cosmo}{List with output from SC-COSMO and all parameters}

\item{proportion}{Logical. Divides incident infections by total population}

\item{age_proportion}{Logical. Divides incident infections by age-specific total population}

\item{only_all}{logical. Should only overall total infections be plotted}
}
\value{
A data.frame with the total number of incident infections by age group and in total
as columns over time
}
\description{
\code{plot_infinc_totals} Plot total number of incident infections by age group
and/or in total over time.
}
