% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_validation_functions.R
\name{plot_model_out_vs_targets}
\alias{plot_model_out_vs_targets}
\title{Plot Targets}
\usage{
plot_model_out_vs_targets(
  df_all_targets,
  df_model_out,
  print_plot = TRUE,
  return_plot = TRUE
)
}
\arguments{
\item{print_plot}{Flag (default is TRUE) of whether the plot should be
printed in screen.}

\item{return_plot}{Flag (default is TRUE) of whether the plot should be
returned.}

\item{l_model_out}{List with decision model outputs.}

\item{l_targets}{List calibration targets.}
}
\value{
A ggplot2 object.
}
\description{
\code{plot_model_out_vs_targets} plots model outputs vs targets.
}
