% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_cumulative_mortality_risk.R
\name{plot_mort_risk}
\alias{plot_mort_risk}
\title{Plot expected cumulative mortality risk}
\usage{
plot_mort_risk(df_CMR, n_MR_day = 30, save_plot = FALSE)
}
\arguments{
\item{df_CMR}{Data.frame. Data of cumulative mortality risk generated by
\code{get_cum_mort_risk}.}

\item{n_MR_day}{Day of expected mortality risk. Options: 30, 60.}

\item{save_plot}{Flag (default is FALSE) of whether to save plot.}
}
\value{
A ggplot object.
}
\description{
\code{plot_mort_risk} plot expected cumulative mortality risk of Covid-19 for MCMA.
}
