% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_model_output_calc_plot_functions.R
\name{plot_popsize_totals}
\alias{plot_popsize_totals}
\title{Plot total population}
\usage{
plot_popsize_totals(l_out_cosmo, only_all = TRUE, print_plot = TRUE)
}
\arguments{
\item{l_out_cosmo}{List with output from SC-COSMO and all parameters.}

\item{only_all}{Flag (default is TRUE) of whether only overall ("All") population
should be plotted.}

\item{print_plot}{Flag (default is TRUE) of whether the plot should be printed
in screen.}
}
\value{
A ggplot object.
}
\description{
\code{plot_popsize_totals} plot total population by age group and/or
in total over time.
}
