% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_model_output_calc_plot_functions.R
\name{plot_rec_totals}
\alias{plot_rec_totals}
\title{Plot total recovered}
\usage{
plot_rec_totals(
  l_out_cosmo,
  proportion = FALSE,
  age_proportion = FALSE,
  only_all = TRUE,
  print_plot = TRUE
)
}
\arguments{
\item{l_out_cosmo}{List with output from SC-COSMO and all parameters.}

\item{proportion}{Flag (default is FALSE) of whether to divide total
number of recovered by total population.}

\item{age_proportion}{Flag (default is FALSE) of whether to divide
total number of recovered by age-specific total population.}

\item{only_all}{logical. Should only overall total infections be plotted}

\item{print_plot}{Flag (default is TRUE) of whether the plot should be printed
in screen.}
}
\value{
A ggplot object.
}
\description{
\code{plot_rec_totals} plot total number of recovered
by age group and/or in total over time.
}
