% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_target_functions.R
\name{plot_targets}
\alias{plot_targets}
\title{Plot Targets}
\usage{
plot_targets(l_targets, print_plot = TRUE, save_plot = TRUE)
}
\arguments{
\item{l_targets}{List with calibration targets.}

\item{print_plot}{Flag (default is FALSE) of whether to print plot.}

\item{save_plot}{Flag (default is FALSE) of whether to save plot.}
}
\value{
A ggplot object.
}
\description{
\code{plot_targets} plots calibration targets.
}
