% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_calibration_functions.R
\name{posterior}
\alias{posterior}
\title{Evaluate posterior of calibrated parameters}
\usage{
posterior(v_params, ...)
}
\arguments{
\item{v_params}{Vector (or matrix) of model parameters.}

\item{...}{Further arguments to be passed to.}
}
\value{
A scalar (or vector) with posterior values.
}
\description{
\code{posterior} computes a posterior value for one (or multiple) parameter
set(s).
}
