% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_hosp_calibration_functions.R
\name{prep_dx_hosp_calib}
\alias{prep_dx_hosp_calib}
\title{Produce hospitalization occupation over time}
\usage{
prep_dx_hosp_calib(l_params_all, use_prevalence = FALSE)
}
\arguments{
\item{use_prevalence}{Flag (default is FALSE) of whether to use prevalence as
input to hospitalization model. If FALSE, incidence is used instead of
prevalence.}

\item{l_out_cosmo}{List with output from SC-COSMO and all parameters.}
}
\value{
A list of incident hospitalizations by type (Tot , Non ICU, or ICU),
age group (including All), and time
and of prevalent hospitalizations by type (Tot , Non ICU, or ICU),
age group (including All), and time.
}
\description{
\code{prep_dx_hosp_calib} computes both incident hospitalizations and prevalence
in hospitalizations (Tot, Non ICU and ICU) by age groups to be used for
calibration routines.
}
