% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_hospital_model_functions.R
\name{prep_dx_hospitalizations}
\alias{prep_dx_hospitalizations}
\title{Produce hospitalization occupation over time}
\usage{
prep_dx_hospitalizations(l_out_cosmo, use_prevalence = FALSE)
}
\arguments{
\item{l_out_cosmo}{List with output from SC-COSMO and all parameters.}

\item{use_prevalence}{Flag (default is FALSE) of whether to use prevalence as
input to hospitalization model. If FALSE, incidence is used instead of
prevalence.}
}
\value{
A list of incident hospitalizations by type (Tot , Non ICU, or ICU),
age group (including All), and time
and of prevalent hospitalizations by type (Tot , Non ICU, or ICU),
age group (including All), and time.
}
\description{
\code{prep_dx_hospitalizations} computes both incident hospitalizations
and prevalence in hospital also by Non ICU and ICU, and age groups.
}
