% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_calibration_functions.R
\name{prior}
\alias{prior}
\title{Evaluate prior of calibrated parameters}
\usage{
prior(v_params)
}
\arguments{
\item{v_params}{Vector (or matrix) of model parameters.}
}
\value{
A scalar (or vector) with prior values.
}
\description{
\code{prior} computes a prior value for one (or multiple) parameter set(s).
}
