% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_projections_functions.R
\name{project_epi_out}
\alias{project_epi_out}
\title{Calculate epidemiological outcomes projections from SC-COSMO}
\usage{
project_epi_out(v_states_project, l_params_all, n_date_ini, n_lag_inf = NULL)
}
\arguments{
\item{v_states_project}{Vector specifying the name of the states to be
projected.}

\item{l_params_all}{List of all SC-COSMO model parameters.}

\item{n_date_ini}{Initial calendar date of the simulation.}

\item{n_lag_inf}{Lag in time series of infectious individuals.}
}
\value{
A list with epidemiological outcomes.
}
\description{
\code{project_epi_out} projects epidemiological outcomes from the
SC-COSMO model for selected states in Mexico.
}
