% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_projections_functions.R
\name{project_interventions_probabilistic}
\alias{project_interventions_probabilistic}
\title{Probabilistic projections of interventions}
\usage{
project_interventions_probabilistic(
  m_calib_post,
  n_date_ini,
  v_n_date0_NPI,
  n_t_calib,
  n_t_project,
  n_lag_inf
)
}
\arguments{
\item{m_calib_post}{Matrix with calibrated parameters from posterior
distribution.}

\item{n_date_ini}{Initial date of calibration.}

\item{v_n_date0_NPI}{Vector with the time steps (\code{0 = date_init}) at
which effect of NPI changed in the calibration period.}

\item{n_t_calib}{Number of calibration days.}

\item{n_t_project}{Number of projection days.}

\item{n_lag_inf}{Lag in time series of infectious individuals.}
}
\value{
A list with probabilistic projections of interventions.
}
\description{
\code{project_interventions_probabilistic} produces probabilistic projections
of interventions
}
