% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_calibration_functions.R
\name{sample.prior}
\alias{sample.prior}
\title{Sample from prior distributions of calibrated parameters}
\usage{
sample.prior(
  n_samp,
  v_param_names = names(get_bounds()$v_lb),
  v_lb = get_bounds()$v_lb,
  v_ub = get_bounds()$v_ub
)
}
\arguments{
\item{n_samp}{Number of samples.}

\item{v_param_names}{Vector with parameter names.}

\item{v_lb}{Vector with upper bounds for each parameter.}

\item{v_ub}{Vector with lower bounds for each parameter.}
}
\value{
A matrix with number of calibrated parameters as columns and \code{n_samp}
rows. Each row corresponds to a parameter set sampled from their prior
distributions.
}
\description{
\code{sample.prior} generates a sample of parameter sets from their prior
distribution.
}
