% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sccosmomcma.R
\docType{package}
\name{sccosmo}
\alias{sccosmo}
\title{\code{sccosmo} package}
\description{
The Stanford-CIDE COronavirus Simulation MOdel (SC-COSMO)
}
\details{
See the description of the package on the README on
\href{https://github.com/feralaes/sccosmo#readme}{GitHub}
}
