% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_validation_functions.R
\name{validate_out}
\alias{validate_out}
\title{Generate outputs for validation}
\usage{
validate_out(
  m_calib_post,
  l_params_all,
  n_date_ini,
  n_t_calib,
  n_lag_inf = 14,
  l_dates_targets
)
}
\arguments{
\item{m_calib_post}{Matrix with calibrated parameters from posterior
distribution}

\item{l_params_all}{List with all parameters of the decision model.}

\item{n_date_ini}{Initial date of calibration.}

\item{n_t_calib}{Number of calibration days.}

\item{n_lag_inf}{Lag in time series of infectious individuals.}

\item{l_dates_targets}{List of initial and last dates of targets series.}
}
\value{
A list with outputs for validation.
}
\description{
\code{validate_out} produces output to validate the model.
}
