% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_contact_matrix.R
\name{wrangle_contact_matrix}
\alias{wrangle_contact_matrix}
\title{Get wrangle contact matrix}
\usage{
wrangle_contact_matrix(
  setting = "all_locations",
  src_country = src_country,
  v_init_age_grps = v_init_age_grps
)
}
\arguments{
\item{setting}{Character that specifies which contact matrix to return.
Options: all_locations, home, non_home, school, work, other_locations.}

\item{src_country}{Character that specifies which country contact matrix to
return.}

\item{v_init_age_grps}{Vector that specifies the age bins to aggregate and
return.}
}
\description{
\code{wrangle_contact_matrix} returns a data.frame with all of the
processed information on population age structure.
}
