% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_time_varying_functions.R
\name{add_period}
\alias{add_period}
\title{Helper function add periods together into a list containing all periods}
\usage{
add_period(l_period_def = NULL, l_period_add)
}
\arguments{
\item{l_period_def}{the existing list of period definitions}

\item{l_period_add}{the period to add to the full list, should be defined using make_period}
}
\value{
List (l_period_def) appended with l_period_add as a new entry
}
\description{
\code{add_period} Adds time periods as new entries in list of time periods
}
