% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_time_varying_functions.R
\name{get_const_multiage_list}
\alias{get_const_multiage_list}
\title{Helper convenience function that defines age-specific lists of periods for values that
will remain constant over the duration of the simulation.}
\usage{
get_const_multiage_list(time_stop, v_ageval)
}
\arguments{
\item{time_stop}{Time period end (numeric, in days) (should be \code{n_t}).}

\item{v_ageval}{Vector of constant values for each each group.}
}
\value{
List (\code{l_l_period}) list of lists of periods (1 per age group)
with constant values.
}
\description{
\code{get_const_multiage_list} adds lists of time periods as new entries in
list of age group lists.
}
