% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_mortality_functions.R
\name{get_ifr}
\alias{get_ifr}
\title{Get the fraction of undetected cases that die}
\usage{
get_ifr(time, l_params_all)
}
\arguments{
\item{time}{Time (numeric, in days) at which \code{v_ifr} is evaluated.}

\item{l_params_all}{List with all parameters of decision model.}
}
\value{
Proportion (age-specific) of undetected cases that die at a given point in time.
}
\description{
\code{get_ifr} pulls age-specific fraction of undetected cases that die at a
given time.
}
