% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_model_inputs_functions.R
\name{load_params_init}
\alias{load_params_init}
\title{Base-case initial parameter set}
\usage{
load_params_init(
  n_t = 30,
  time_step = 1,
  comp = TRUE,
  v_init_age_grps = c(0, 5, 15, 25, 45, 55, 65, 70),
  v_inf_init_ages = NULL,
  ctry = "Mexico",
  ste = "Mexico City",
  cty = "Mexico City",
  l_contact_info = l_contact_matrices,
  r_birth = 0,
  r_beta = 0.0536,
  v_sigma = rep(1/3, length(v_init_age_grps)),
  v_gamma = rep(1/3.1, length(v_init_age_grps)),
  v_omega = rep(0, length(v_init_age_grps)),
  l_nu_exp1_dx = add_period(l_period_def = NULL, l_period_add =
    make_period(functional_form = "constant", time_start = 0, time_stop = n_t, val_start
    = 0, val_end = 0)),
  l_nu_exp2_dx = add_period(l_period_def = NULL, l_period_add =
    make_period(functional_form = "constant", time_start = 0, time_stop = n_t, val_start
    = 1/12, val_end = 1/12)),
  l_nu_inf1_dx = add_period(l_period_def = NULL, l_period_add =
    make_period(functional_form = "constant", time_start = 0, time_stop = n_t, val_start
    = 0, val_end = 0)),
  l_nu_inf2_dx = add_period(l_period_def = NULL, l_period_add =
    make_period(functional_form = "constant", time_start = 0, time_stop = n_t, val_start
    = 1/12, val_end = 1/12)),
  l_phi_exp1_dx = add_period(l_period_def = NULL, l_period_add =
    make_period(functional_form = "constant", time_start = 0, time_stop = n_t, val_start
    = 0, val_end = 0)),
  l_phi_exp2_dx = add_period(l_period_def = NULL, l_period_add =
    make_period(functional_form = "constant", time_start = 0, time_stop = n_t, val_start
    = 0, val_end = 0)),
  l_phi_inf1_dx = add_period(l_period_def = NULL, l_period_add =
    make_period(functional_form = "constant", time_start = 0, time_stop = n_t, val_start
    = 0, val_end = 0)),
  l_phi_inf2_dx = add_period(l_period_def = NULL, l_period_add =
    make_period(functional_form = "constant", time_start = 0, time_stop = n_t, val_start
    = 0, val_end = 0)),
  l_cfr = get_const_multiage_list(n_t, c(1.3e-05, 8.7e-05, 0.000374, 0.00146, 0.007725,
    0.0262, 0.05145, 0.134)),
  l_ifr = get_const_multiage_list(n_t, c(8.05e-06, 4.28e-05, 0.00018925, 0.000844,
    0.00378, 0.012625, 0.025175, 0.078)),
  l_interventions = add_intervention(interventions = NULL, intervention =
    make_intervention(intervention_type = "StatusQuo", time_start = 0, time_stop = n_t)),
  l_idx_scale_factor = get_const_multiage_list(n_t, rep(0.2, 8)),
  v_reduced_sus = rep(1, 8),
  v_alpha_dx = NULL,
  l_p_hosp = get_const_multiage_list(n_t, c(1e-05, 2e-04, 0.005, 0.035, 0.06, 0.095,
    0.13, 0.17)),
  v_p_s_hosp = c(0.1111, 0.1111, 0.1202, 0.1202, (0.1202 + 0.1746)/2, (0.1746 +
    0.2875)/2, 0.2875, 0.2875),
  v_p_icu_s_hosp = c(0.191, 0.191, 0.191, 0.191, 0.191, 0.191, 0.191, 0.191),
  v_p_icu_ns_hosp = c(0.024, 0.024, 0.024, 0.024, 0.024, 0.024, 0.024, 0.024),
  m_r_exit_tot = c(1/3.1, 1/3.1, 1/((3.1 + 4.2)/2), 1/4.2, 1/4.3, 1/4.3, 1/4.6, 1/4.6),
  m_r_exit_icu = c(1/7.7, 1/7.7, 1/((7.7 + 7.2)/2), 1/7.2, 1/7.3, 1/7.3, 1/7.1, 1/7.1),
  m_r_exit_nonicu = 0.5 * m_r_exit_tot + 0.5 * m_r_exit_icu,
  m_sigma_tot = c(2, 2, 2, 2, 2, 2, 2, 2),
  m_sigma_nonicu = c(2, 2, 2, 2, 2, 2, 2, 2),
  m_sigma_icu = c(1.5, 1.5, 1.5, 1.5, 1.5, 1.5, 1.5, 1.5),
  n_hhsize = 3,
  r_tau = 0,
  r_sigma = v_sigma[1],
  r_gamma = v_gamma[1],
  r_omega = v_omega[1]
)
}
\arguments{
\item{n_t}{Time horizon in days.}

\item{time_step}{Model evaluations per day.}

\item{comp}{Flag (default TRUE) of whether the compiled version of SC-COSMO
should be used.}

\item{v_init_age_grps}{Vector with initial ages of age groups.}

\item{v_inf_init_ages}{Vector with number of individuals in first
infectious class for each age group. Default will assign 1 for each age
group.}

\item{ctry}{Country.}

\item{ste}{State.}

\item{cty}{County or municipality.}

\item{l_contact_info}{A list with overall and setting specific contact
matrices.}

\item{r_birth}{Daily birth rate (crude).}

\item{r_beta}{Transmission parameters.}

\item{v_sigma}{Age-specific daily rate of progression of exposed individuals
(Latent period).}

\item{v_gamma}{Age-specific daily rate of recovery of infectious individuals
(Infectiousness period).}

\item{v_omega}{Age-specific daily rate of waning immunity of recovered
individuals.}

\item{l_nu_exp1_dx}{List specifying time-varying case detection when exposed
\code{l = 1}.}

\item{l_nu_exp2_dx}{List specifying time-varying case detection when exposed
\code{l = 2}.}

\item{l_nu_inf1_dx}{List specifying time-varying case detection when infectious
\code{l = 1}.}

\item{l_nu_inf2_dx}{List specifying time-varying case detection when infectious
\code{l = 2}.}

\item{l_phi_exp1_dx}{List specifying time-varying screen detection when exposed
\code{l = 1}.}

\item{l_phi_exp2_dx}{List specifying time-varying screen detection when exposed
\code{l = 2}.}

\item{l_phi_inf1_dx}{List specifying time-varying screen detection when
infectious \code{l = 1}.}

\item{l_phi_inf2_dx}{List specifying time-varying screen detection when
infectious \code{l = 2}.}

\item{l_cfr}{Age-specific case fatality rate.}

\item{l_ifr}{Age-specific infection fatality rate.}

\item{l_interventions}{List of intervention "objects" defining type, timing,
and effects of interventions.}

\item{l_idx_scale_factor}{Age-specific infectiousness reduction factor on
detected cases.}

\item{v_reduced_sus}{Age-specific reduction factor on susceptibility.}

\item{v_alpha_dx}{Age-specific reduction in mortality on detected infectious
vs undetected infectious.}

\item{l_p_hosp}{Age-specific probability of hospitalization for detected (prevalent or
incident) cases.}

\item{v_p_s_hosp}{Age-specific proportion of hospitalizations that are
severe based on NEJM definition.}

\item{v_p_icu_s_hosp}{Age-specific probability of going to ICU given that
you are severe hospitalization.}

\item{v_p_icu_ns_hosp}{Age-specific probability of going to ICU given that
you are non-severe hospitalization.}

\item{m_r_exit_tot}{Age-specific rate of non-severe hospital exit}

\item{m_r_exit_icu}{Age-specific rate of ICU hospital exit}

\item{m_r_exit_nonicu}{Age-specific rate of severe hospital exit}

\item{m_sigma_tot}{Age-specific sigma (gamma distribution) total
hospitalization duration.}

\item{m_sigma_nonicu}{Age-specific sigma (gamma distribution) non ICU
hospitalizations duration.}

\item{m_sigma_icu}{Age-specific sigma (gamma distribution) ICU hospitalization
duration.}

\item{n_hhsize}{Household size (integer).}

\item{r_tau}{Transmission rate in the household model.}

\item{r_sigma}{Daily rate of progression of exposed individuals
(Latent period). Default to the first entry of the age-specific \code{v_sigma}.}

\item{r_gamma}{Daily rate of recovery of infectious individuals
(Infectiousness period). Default to the first entry of the age-specific
\code{v_gamma}.}

\item{r_omega}{Waning rate for the household model (default to the first
entry of the age-specific \code{v_omega}).
Default will assign a reduction factor of 1 (i.e., no social distancing).}
}
\value{
List of all parameters
}
\description{
\code{load_params_init} generates the initial values of the SC-COSMO parameters.
}
