% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_time_varying_functions.R
\name{make_period}
\alias{make_period}
\title{Helper function to define a time chunk for a time varying parameter}
\usage{
make_period(functional_form, time_start, time_stop, val_start, val_end, ...)
}
\arguments{
\item{functional_form}{Character string of "linear", "constant", or
"general_logit".}

\item{time_start}{Numeric time period start (in days).}

\item{time_stop}{Numeric time period end (in days).}

\item{val_start}{Starting value for time varying parameter at \code{time_start}.}

\item{val_end}{Ending value for time varying parameter at \code{time_end}.}

\item{...}{Further arguments to be passed to.}
}
\value{
A list with parameters initialized.
}
\description{
\code{make_period} defines a time chunk for a generic time varying parameter.
Includes some required arguments, but also allows additional arguments.
}
