% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_model_output_calc_plot_functions.R
\name{plot_cosmo_data}
\alias{plot_cosmo_data}
\title{Plot COSMO data}
\usage{
plot_cosmo_data(
  df_plotting,
  proportion = FALSE,
  age_proportion = FALSE,
  only_all = TRUE,
  print_plot = TRUE,
  label_y
)
}
\arguments{
\item{df_plotting}{data frame of data to be plotted}

\item{proportion}{Logical. Divides cumulative infections by total population}

\item{age_proportion}{Logical. Divides cumulative infections by age-specific total population}

\item{only_all}{logical. Should only overall total infections be plotted}

\item{print_plot}{logical. Should show it?}
}
\value{
A ggplot object
}
\description{
\code{plot_cosmo_data} A generic function called by various wrapper
funcitons to plot a data frame returned by one of our calc methods by
age group and/or in total over time.
}
