% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_model_output_calc_plot_functions.R
\name{plot_incdeathsdx_totals}
\alias{plot_incdeathsdx_totals}
\title{Plot total Incident COVID19 deaths from diagnosed infections}
\usage{
plot_incdeathsdx_totals(
  l_out_cosmo,
  proportion = FALSE,
  age_proportion = FALSE,
  only_all = TRUE,
  print_plot = TRUE
)
}
\arguments{
\item{l_out_cosmo}{List with output from SC-COSMO and all parameters}

\item{only_all}{logical. Should only overall total deaths from COVID19 be plotted}
}
\value{
A data.frame with the total number of COVID19 deaths from diagnosed
infections by age group and in total as columns over time
}
\description{
\code{plot_incdeathsdx_totals} Plot total incident number of COVID19
deaths from diagnosed infections by age group and/or in total over time.
}
