% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_model_output_calc_plot_functions.R
\name{plot_infinc_totals}
\alias{plot_infinc_totals}
\title{Plot total incident infections}
\usage{
plot_infinc_totals(
  l_out_cosmo,
  proportion = FALSE,
  age_proportion = FALSE,
  only_all = TRUE,
  print_plot = TRUE
)
}
\arguments{
\item{l_out_cosmo}{List with output from SC-COSMO and all parameters.}

\item{proportion}{Flag (default is FALSE) of whether to divide total number of
incident infections by total population.}

\item{age_proportion}{Flag (default is FALSE) of whether to divide
total number of incident infections by age-specific total population.}

\item{only_all}{Flag (default is TRUE) of whether only overall ("All") total
number of incident infections should be plotted.}

\item{print_plot}{Flag (default is TRUE) of whether the plot should be printed
in screen.}
}
\value{
A ggplot object.
}
\description{
\code{plot_infinc_totals} plot total number of incident infections by age group
and/or in total over time.
}
