% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_model_output_calc_plot_functions.R
\name{plot_popsize_totals}
\alias{plot_popsize_totals}
\title{Plot total population}
\usage{
plot_popsize_totals(l_out_cosmo, only_all = TRUE, print_plot = TRUE)
}
\arguments{
\item{l_out_cosmo}{List with output from SC-COSMO and all parameters}

\item{only_all}{logical. Should only overall total infections be plotted}
}
\value{
A data.frame with the total population by age group and in total
as columns over time
}
\description{
\code{plot_popsize_totals} Plot total population by age group and/or
in total over time.
}
