% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_hospital_model_functions.R
\name{prep_dx_hospitalizations}
\alias{prep_dx_hospitalizations}
\title{Produce hospitalization occupation over time}
\usage{
prep_dx_hospitalizations(l_out_cosmo, use_prevalence = FALSE)
}
\arguments{
\item{l_out_cosmo}{List with output from SC-COSMO and all parameters}

\item{use_prevalence}{Indicator to use prevalence as input to hospitalization
model. If FALSE, incidence is used instead of prevalence (Default == FALSE).}
}
\value{
A list of incident hosptalizations by type (Not Severe , Severe, or ICU),
age group (including All), and time
and of prevalent hosptalizations by type (Not Severe , Severe, or ICU),
age group (including All), and time
}
\description{
Computes both incident hospitalizations and prevalence in hospital
also by hospitalized not severe, severe and ICU, and age groups.
}
